+++ src/Symfony/Bridge/Doctrine/CHANGELOG.md
@@ -5,7 +5,7 @@ CHANGELOG
 -----
 
  * added `DoctrineClearEntityManagerWorkerSubscriber`
- * deprecated `RegistryInterface`, use `Doctrine\Common\Persistence\ManagerRegistry`
+ * deprecated `RegistryInterface`, use `Doctrine\Persistence\ManagerRegistry`
  * added support for invokable event listeners
  * added `getMetadataDriverClass` method to deprecate class parameters in service configuration files